First go at new item group api
