+++ fabric-item-group-api-v1/build.gradle
@@ -1,0 +1,7 @@
+archivesBaseName = "fabric-item-group-api-v1"
+version = getSubprojectVersion(project)
+
+moduleDependencies(project, [
+		'fabric-api-base',
+		'fabric-resource-loader-v0'
+])
