+++ fabric-item-group-api-v1/src/client/java/net/fabricmc/fabric/mixin/itemgroup/client/CreativeInventoryScreenMixin.java
@@ -1,0 +1,168 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.mixin.itemgroup.client;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import net.minecraft.class_7706;
+import net.minecraft.client.gui.screen.ingame.AbstractInventoryScreen;
+import net.minecraft.client.gui.screen.ingame.CreativeInventoryScreen;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.item.ItemGroup;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.text.Text;
+
+import net.fabricmc.fabric.impl.itemgroup.CreativeGuiExtensions;
+import net.fabricmc.fabric.impl.itemgroup.FabricCreativeGuiComponents;
+
+@Mixin(CreativeInventoryScreen.class)
+public abstract class CreativeInventoryScreenMixin<T extends ScreenHandler> extends AbstractInventoryScreen<T> implements CreativeGuiExtensions {
+	public CreativeInventoryScreenMixin(T screenHandler, PlayerInventory playerInventory, Text text) {
+		super(screenHandler, playerInventory, text);
+	}
+
+	@Shadow
+	protected abstract void setSelectedTab(ItemGroup itemGroup_1);
+
+	@Shadow
+	public abstract int getSelectedTab(); /* XXX getSelectedTab XXX */
+
+	// "static" matches selectedTab
+	private static int fabric_currentPage = 0;
+
+	private int fabric_getPageOffset(int page) {
+		return switch (page) {
+		case 0 -> 0;
+		case 1 -> 12;
+		default -> 12 + ((12 - FabricCreativeGuiComponents.COMMON_GROUPS.size()) * (page - 1));
+		};
+	}
+
+	private int fabric_getOffsetPage(int offset) {
+		if (offset < 12) {
+			return 0;
+		} else {
+			return 1 + ((offset - 12) / (12 - FabricCreativeGuiComponents.COMMON_GROUPS.size()));
+		}
+	}
+
+	@Override
+	public void fabric_nextPage() {
+		if (fabric_getPageOffset(fabric_currentPage + 1) >= class_7706.field_40207.length) {
+			return;
+		}
+
+		fabric_currentPage++;
+		fabric_updateSelection();
+	}
+
+	@Override
+	public void fabric_previousPage() {
+		if (fabric_currentPage == 0) {
+			return;
+		}
+
+		fabric_currentPage--;
+		fabric_updateSelection();
+	}
+
+	@Override
+	public boolean fabric_isButtonVisible(FabricCreativeGuiComponents.Type type) {
+		return class_7706.field_40207.length > 12;
+	}
+
+	@Override
+	public boolean fabric_isButtonEnabled(FabricCreativeGuiComponents.Type type) {
+		if (type == FabricCreativeGuiComponents.Type.NEXT) {
+			return !(fabric_getPageOffset(fabric_currentPage + 1) >= class_7706.field_40207.length);
+		}
+
+		if (type == FabricCreativeGuiComponents.Type.PREVIOUS) {
+			return fabric_currentPage != 0;
+		}
+
+		return false;
+	}
+
+	private void fabric_updateSelection() {
+		int minPos = fabric_getPageOffset(fabric_currentPage);
+		int maxPos = fabric_getPageOffset(fabric_currentPage + 1) - 1;
+		int curPos = getSelectedTab();
+
+		if (curPos < minPos || curPos > maxPos) {
+			setSelectedTab(class_7706.field_40207[fabric_getPageOffset(fabric_currentPage)]);
+		}
+	}
+
+	@Inject(method = "init", at = @At("RETURN"))
+	private void init(CallbackInfo info) {
+		fabric_updateSelection();
+
+		int xpos = x + 116;
+		int ypos = y - 10;
+
+		addDrawableChild(new FabricCreativeGuiComponents.ItemGroupButtonWidget(xpos + 11, ypos, FabricCreativeGuiComponents.Type.NEXT, this));
+		addDrawableChild(new FabricCreativeGuiComponents.ItemGroupButtonWidget(xpos, ypos, FabricCreativeGuiComponents.Type.PREVIOUS, this));
+	}
+
+	@Inject(method = "setSelectedTab", at = @At("HEAD"), cancellable = true)
+	private void setSelectedTab(ItemGroup itemGroup, CallbackInfo info) {
+		if (!fabric_isGroupVisible(itemGroup)) {
+			info.cancel();
+		}
+	}
+
+	@Inject(method = "renderTabTooltipIfHovered", at = @At("HEAD"), cancellable = true)
+	private void renderTabTooltipIfHovered(MatrixStack matrixStack, ItemGroup itemGroup, int mx, int my, CallbackInfoReturnable<Boolean> info) {
+		if (!fabric_isGroupVisible(itemGroup)) {
+			info.setReturnValue(false);
+		}
+	}
+
+	@Inject(method = "isClickInTab", at = @At("HEAD"), cancellable = true)
+	private void isClickInTab(ItemGroup itemGroup, double mx, double my, CallbackInfoReturnable<Boolean> info) {
+		if (!fabric_isGroupVisible(itemGroup)) {
+			info.setReturnValue(false);
+		}
+	}
+
+	@Inject(method = "renderTabIcon", at = @At("HEAD"), cancellable = true)
+	private void renderTabIcon(MatrixStack matrixStack, ItemGroup itemGroup, CallbackInfo info) {
+		if (!fabric_isGroupVisible(itemGroup)) {
+			info.cancel();
+		}
+	}
+
+	private boolean fabric_isGroupVisible(ItemGroup itemGroup) {
+		if (FabricCreativeGuiComponents.COMMON_GROUPS.contains(itemGroup)) {
+			return true;
+		}
+
+		return fabric_currentPage == fabric_getOffsetPage(itemGroup.getIndex());
+	}
+
+	@Override
+	public int fabric_currentPage() {
+		return fabric_currentPage;
+	}
+}
