+++ fabric-item-group-api-v1/src/client/resources/fabric-item-groups-v1.client.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": true,
+  "package": "net.fabricmc.fabric.mixin.itemgroup.client",
+  "compatibilityLevel": "JAVA_16",
+  "client": [
+    "ItemGroupMixin",
+    "CreativeInventoryScreenMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
