+++ fabric-item-group-api-v1/src/main/resources/assets/fabric-item-groups-v0/icon.png
