+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/de_de.json
