+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/es_es.json
