+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/fa_ir.json
