+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/fr_fr.json
