+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/ko_kr.json
