+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/pt_br.json
