+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/lang/sv_se.json
