+++ fabric-item-group-api-v1/src/main/resources/assets/fabric/textures/gui/creative_buttons.png
