+++ fabric-item-group-api-v1/src/main/resources/fabric.mod.json
@@ -1,0 +1,34 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-item-groups-v1",
+  "name": "Fabric Item Groups (v1)",
+  "version": "${version}",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-item-groups-v0/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.14.9",
+    "fabric-api-base": "*",
+    "fabric-resource-loader-v0": "*"
+  },
+  "description": "An API for adding custom item groups.",
+  "mixins": [
+    "fabric-item-groups-v1.mixins.json",
+    {
+      "config": "fabric-item-groups-v1.client.mixins.json",
+      "environment": "client"
+    }
+  ],
+  "custom": {
+    "fabric-api:module-lifecycle": "stable"
+  }
+}
