--- fabric-item-groups-v0/build.gradle
@@ -1,7 +1,0 @@
-archivesBaseName = "fabric-item-groups-v0"
-version = getSubprojectVersion(project)
-
-moduleDependencies(project, [
-		'fabric-api-base',
-		'fabric-resource-loader-v0'
-])
