--- fabric-item-groups-v0/src/client/resources/fabric-item-groups-v0.client.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.item.group.client",
-  "compatibilityLevel": "JAVA_16",
-  "client": [
-    "ItemGroupMixin",
-    "CreativeInventoryScreenMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
