--- fabric-item-groups-v0/src/main/resources/fabric-item-groups-v0.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.item.group",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-    "ItemGroupMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
