--- fabric-item-groups-v0/src/main/resources/fabric.mod.json
@@ -1,35 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-item-groups-v0",
-  "name": "Fabric Item Groups (v0)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-item-groups-v0/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.6.0",
-    "minecraft": ">=1.15-",
-    "fabric-api-base": "*",
-    "fabric-resource-loader-v0": "*"
-  },
-  "description": "An API for adding custom item groups.",
-  "mixins": [
-    "fabric-item-groups-v0.mixins.json",
-    {
-      "config": "fabric-item-groups-v0.client.mixins.json",
-      "environment": "client"
-    }
-  ],
-  "custom": {
-    "fabric-api:module-lifecycle": "stable"
-  }
-}
