+++ gradle.properties
@@ -4,7 +4,7 @@ org.gradle.parallel=true
 version=0.64.0
 minecraft_version=22w42a
 yarn_version=+build.2
-loader_version=0.14.9+local
+loader_version=0.14.9
 
 prerelease=true
 
@@ -27,7 +27,7 @@ fabric-events-lifecycle-v0-version=0.2.29
 fabric-game-rule-api-v1-version=1.0.22
 fabric-gametest-api-v1-version=1.1.1
 fabric-item-api-v1-version=1.5.8
-fabric-item-groups-v0-version=0.3.32
+fabric-item-group-api-v1-version=1.0.0
 fabric-key-binding-api-v1-version=1.0.22
 fabric-keybindings-v0-version=0.2.20
 fabric-lifecycle-events-v1-version=2.2.0
