+++ settings.gradle
@@ -26,7 +26,7 @@ include 'fabric-crash-report-info-v1'
 //include 'fabric-game-rule-api-v1'
 //include 'fabric-gametest-api-v1'
 //include 'fabric-item-api-v1'
-//include 'fabric-item-groups-v0'
+include 'fabric-item-group-api-v1'
 include 'fabric-key-binding-api-v1'
 include 'fabric-lifecycle-events-v1'
 //include 'fabric-loot-api-v2'
